change type of w and h to uint8_t, as it is stored and read as a single byte from the file

+++ maze.cpp
@@ -79,7 +79,7 @@ Node::operator std::string()
 ***************************************/
 #pragma region Maze
 
-Maze::Maze(uint _w, uint _h)
+Maze::Maze(uint8_t _w, uint8_t _h)
 {
     w = _w;
     h = _h;
